iota_proto_build::codegen

Module accessor_config

Structs

AccessorTypes

Flags for different types of accessor methods to generate

Functions

parse_proto_accessors_from_pool

Parse proto files to extract message_accessors and field_accessors annotations from the descriptor pool Returns a map of “MessageName.field_name” -> AccessorTypes

Type Aliases

AccessorMap

Map of field names to their accessor configurations Key is “message_name.field_name”, value is the accessor types